tdf#91762: Check wether the row and the cell have text...
... in order to set the row's min height

Regression from 5cb58d045c83a9e11dfd850b124405835b0c7fc3
After this commit, the row's min height was set based only on
the cells containing text in the row, but the problem appeared
when the row didn't have any cell with text.
Change logic to check wether there's text in the cell and in the row.

Now, height in SdImportTest::testRowHeight() is 507 instead of
508 but I can't figure it out why. However, I believe there's
no harm in change the test from 508 to 507 as, visually speaking,
the difference can't be distinguish.

Change-Id: I0b3a14c34eaeaa8e77227860ca290fb79a0302ce
Reviewed-on: https://gerrit.libreoffice.org/21692
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: jan iversen <jani@documentfoundation.org>