sw tiled rendering: avoid unnecessary invalidation in SwView::SetVisArea()
SwWrtShell's visible area is set to the entire document since
8bf83f3d3cec75b95c67ad14bd9d6940a2bef449 (tdf#94237 tiled rendering: Use
the entire document as the visual area., 2015-09-15).

Let's be consistent and do the same for SwView, so that
SwView::PageDown() and all other similar functions do not cause
unnecessary invalidations, as this way later we'll realize that SwView's
and SwWrtShell's visible area is the same.

Change-Id: Ia22f07ddfb18c6f5ab6cbafede7cf8799b1177a1