vcl: move vcl/inc/list.h to vcl/source/fontsubset/
This appears to be an implementation of a linked list for processing
truetype fonts. I genuinely do NOT want anyone to use this anywhere
in the codebase except where it's already being used! If we must use
a list, then we can use the standard template library...

Thus, I'm moving this closer to the code that uses it - which means
moving it directly into the fontsubset directory

Change-Id: I67fb324a004cd30e32043bf9d5a3b193351d58c0