vcl: remove vcl/inc/magic.h
I must have a sensitive olfactory, but this is pretty smelly. One
file that has a single define, which is a "MAX_FALLBACK", but is
used in layout *and* glyph fallback code?

Yeah, so probably not that bad, but still not very good. I'm not a
fan of "magic" code.

Change-Id: I8b7f8f060da1b211177af49be20faa8b574ac3f3