libxmlsec: initial OOXML RelationshipTransform
Replace the canonicalization with the steps of actions required by ECMA 376,
part 2, pages 49 - 70:

- parse arguments of the transformation, a SourceId whitelist
- add missing TargetMode attributes

The largest part is to actually keep the data unchanged, everything still needs
to be printed, as the source is a parsed XML tree, while the output is a byte
buffer.

With this, the first _rels/.rels stream of the OOXML document validates for a
test document that is supposed to be valid.

Change-Id: Ie996d93de6a7611bac18a8c37c575363552fdab4
Reviewed-on: https://gerrit.libreoffice.org/21832
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>