xmlsecurity: fix crash on closing DigitalSignaturesDialog without ...
... a signature stream. ODF has a signature stream, OOXML has a
signature storage instead.

Code tries to export all signatures on closing the dialog, but OOXML at
the moment has only an importer; so just fail gracefully in case we only
have a storage, not a stream.

With this, closing the dialog no longer results in a crash for OOXML
documents.

Change-Id: I3e6580ddb20ec31fb0c5a4abe1fab3d8046d0ee5