vcl: move CommandEvent functions from header to cxx file
Moving functions from commandevent.hxx to commandevent.cxx, this means
they are no longer inline, but there was no clear reason as to why they
were inline in the first place!

I've also reordered the header to ensure that CommandEvent is the first
fully declared class, the data classes get declared after this (hence the
forward declarations). This makes reading this header much easier.

One more small thing: a few if statements had braces in the primary if
block, but not in the else block, which is IMO bad style and can possibly
lead to errors. I've made the else block use braces.

Change-Id: I79b010bd66d3c9f4a74d4fa470f62d0263c811f4