tdf#97171: Use DirectWrite for OpenGL glyph caching
This is a squash of several separate intermediate commits, the most
relevant log messages of which are repeated here:

Refactor some previously private methods into public ones and
reimplement parts of D2DWriteTextOutRenderer to user them. Also
apply them to rendering the OpenGL glyph atlas, in an effort to
workaround some bugs in legacy Windows text rendering APIs.

I assume we want to initialise the rectangle with zero
left/right/top/bottom and not using the default constructor, which
sets the right and bottom coordinates to the magic value -32767. That
made the 'bound' rectangle end up with rather amusing boundaries, like
left=-32766, top=-16, right=-32576, bottom=6.

Try calculating a chunks ascent & height from the inkboxes rather than
using the font metrics which might not alway be correct when glyphs
inkboxes are tall than the ascent or lower than the descent.

Mark the mnAscent in the chunk bitmap debug output.

Fix several miscalulations in positioning glyph to be rendered into
that atlas.

Fix vertical alignment problems. Inkboxes are returned with all
co-ordinates relative to the glyphs not the fonts ascent.  Therefor
bounds.Top() is not the vertical overhang but the -ve height of the
inkbox above the baseline.  This fixes the calulation of the per Chunk
ascent.

Fix horizontal occsional alingment issues in OpenGL cached glyphs. The
left edge of the src location rectangle for the first glyph in a cache
chunk would set to extraspace and not zero, but all other rectangles
in the chunk would be set from the aEnds array. This produced a bug
where only certain letters would be mispositioned, proportional to the
fonts point size.

Rename OpenGLGlyphChunk::mnAscent to mnBaselineOffset to reflect curr
use.  Changed at Tor's stuggestion to better describe to it's use as
it's value would be per chunk and based on the maximum ink box bounds
of the glyphs in the chunk, rather than having anything to do with the
font's real ascent value.