sfx2 classification: don't sort items on the toolbar listbox
Assuming the policy follows a least-confidential -> most-confidential
order, doing a lexicographical sort on the category names does more harm
than good. So use a vector instead of a name-indexed map for categories.

Also, don't duplicate the label map for the current state, but just
store the state in the same category structure that is used for other
(currently not used) categories as well.

Change-Id: I1672192e572abfc22b6aeeb152ee7484086cea91