tdf#98224: endless loop in replace all stopped
If a drawing in the document contains the search text then the selection of that
object removes a selection (SwShellCursor) that marks the end of the search.
This is now fixed for that special case.
Replacing text in drawing doesn't work - as before and also not all
occurrences of the searched text are found.

Change-Id: I4a10ddf2dc547572ebb55b6355439d6e320bc980
Reviewed-on: https://gerrit.libreoffice.org/22851
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>
Tested-by: Jenkins <ci@libreoffice.org>